Coroutine Guide Basic - update description (Basic2)
