--- src/main/kotlin/base/Basic5.kt
@@ -1,2 +1,0 @@
-package base
-
