+++ src/main/kotlin/basic/Basic5.kt
@@ -1,0 +1,2 @@
+package basic
+
